Changed Database Ids to non-identity
All the Dust entities have Ids already. I changed the DbContext so that
it generated tables with non-identity keys so that it will use the ones
defined in Dust.
I also updated the logic so that it will search if an Entity exists
already and links to it, rather than creating a new one. This should
only take a single DB hit once per entity, as after that it is cached in
the DbContext.