Local battle times
It took 12 hours to figure out, but I can finally display UTC times in
local time. It took a combination of outputting the DateTime in a
special JS format (2013/1/1 0900 UTC). And then using a JS function to
find all the <time> elements and convert them from that UTC format into
the current locale's time.