Bringing CREST Parser inline
I tried to have the CREST parser as a separate GitHub project. However
the Azure auto deployment fails as it can't bring in the 2nd project.
Long term I'll create a NuGet package and split it back out into a
separate project. But for the moment it is easier to just keep it all in
one repository.